Split Package in various target in order to simplify not using Alamofire
