+++ Templates/Swift/Sources/Models/Enum.swift
