+++ Templates/Swift/Sources/SharedCode/APIClientError.swift
