+++ Templates/Swift/Sources/SharedCode/APIResult.swift
